reef/mako-multiagent/requirements.txt
openai>=1.40.0
kubernetes>=29.0.0
requests>=2.31.0
urllib3>=2.2.0
beautifulsoup4>=4.12.3
numpy>=1.26.0
scikit-learn>=1.3.2
tabulate>=0.9.0
python-dotenv>=1.0.0